blatantly lie about the version here
this branch is to put into the pedant.gemspec so that the pedant tests
against chef-client can use this ipv6-aware rest-client version while
we can still pin to rest-client <= 1.6.7 in the chef gemspec itself.

this should not be used anywhere else outside of the chef-client pedant
gemspec itself.

this stupid hack goes away with chef 12 because it drops the rest-client
dependency completely.

i think chef 11 could drop the dependency as well just as long as
people get a heads up that its being dropped.

even if we got the upstream to take the ipv6 patches, at the point
they'd need to drop them into a new version 1.7.x which wouldn't satisfy
our requirements anyway and would break our builds via the rdoc
dependency, so doing "the right thing" still wouldn't help us at all.

https://i.imgur.com/r7EQEZJ.gif